Add oracle + math finish borrow w/o clockwork

+++ programs/uxd/src/utils/validate_loan_to_value_bps.rs
@@ -2,8 +2,8 @@ use crate::error::UxdError;
 use anchor_lang::prelude::*;
 
 pub fn validate_loan_to_value_bps(
-    loan_to_value_bps: u8,
-    depository_loan_to_value_bps: u8,
+    loan_to_value_bps: u16,
+    depository_loan_to_value_bps: u16,
 ) -> Result<()> {
     require!(loan_to_value_bps > 0, UxdError::InvalidLtvBps);
     require!(
